Users error handler (#9)
* Create an error handler on get all.

* Create an error handler on get one.

* Create an error handler on create.

* Create error handlers on delete.

+++ src/controller/UserController.ts
@@ -7,22 +7,51 @@ export class UserController {
     private userRepository = getRepository(User);
 
     async all(request: Request, response: Response, next: NextFunction) {
-        return this.userRepository.find();
+        const foundAll = await this.userRepository.find();
+
+        if (!foundAll) {
+            response.status(404);
+            return { message: 'Users not found.' };
+        }
+        return foundAll;
     }
 
     async one(request: Request, response: Response, next: NextFunction) {
-        return this.userRepository.findOne(request.params.id);
+        const foundOne = await this.userRepository.findOne(request.params.id);
+
+        if (!foundOne) {
+            response.status(404);
+            return { message: 'User not found.' };
+        }
+        return foundOne;
     }
 
     async create(request: Request, response: Response, next: NextFunction) {
+        if (!request.body.username || !request.body.password) {
+            response.status(422);
+            return { message: "Username and password are required." };
+        }
+
         const userObject = this.userRepository.create(request.body);
         const user = await this.userRepository.save(userObject);
+
+        if (!user) {
+            response.status(500);
+            return { message: 'User could not be created.' };
+        }
+
         return { data: { user } };
     }
 
     async remove(request: Request, response: Response, next: NextFunction) {
-        let userToRemove = await this.userRepository.findOne(request.params.id);
-        return await this.userRepository.remove(userToRemove);
+        const userToRemove = await this.userRepository.findOne(request.params.id);
+
+        if (!userToRemove) {
+            response.status(404);
+            return { message: 'User not found.' };
+        }
+        await this.userRepository.remove(userToRemove);
+        return { message: 'Deleted.' };
     }
 
 }
